restructured Test classes.
