+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/MemberDao.java
@@ -1,4 +1,4 @@
-package com.nashss.se.exchange.dynamodb;
+package com.nashss.se.exchange;
 
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
