+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/activity/CreateItemActivity.java
@@ -4,9 +4,9 @@ import com.nashss.se.exchange.Models.ItemModel;
 import com.nashss.se.exchange.activity.requests.CreateItemRequest;
 import com.nashss.se.exchange.activity.results.CreateItemResult;
 import com.nashss.se.exchange.converters.ModelConverter;
-import com.nashss.se.exchange.dynamodb.Item;
-import com.nashss.se.exchange.dynamodb.ItemDao;
-import com.nashss.se.exchange.dynamodb.MemberDao;
+import com.nashss.se.exchange.Item;
+import com.nashss.se.exchange.ItemDao;
+import com.nashss.se.exchange.MemberDao;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
