+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/activity/CreateMemberActivity.java
@@ -4,8 +4,8 @@ import com.nashss.se.exchange.Models.MemberModel;
 import com.nashss.se.exchange.activity.requests.CreateMemberRequest;
 import com.nashss.se.exchange.activity.results.CreateMemberResult;
 import com.nashss.se.exchange.converters.ModelConverter;
-import com.nashss.se.exchange.dynamodb.Member;
-import com.nashss.se.exchange.dynamodb.MemberDao;
+import com.nashss.se.exchange.Member;
+import com.nashss.se.exchange.MemberDao;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
