+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/activity/GetMemberItemsActivity.java
@@ -4,10 +4,10 @@ import com.nashss.se.exchange.Models.ItemModel;
 import com.nashss.se.exchange.activity.requests.GetMemberItemsRequest;
 import com.nashss.se.exchange.activity.results.GetMemberItemsResult;
 import com.nashss.se.exchange.converters.ModelConverter;
-import com.nashss.se.exchange.dynamodb.Item;
-import com.nashss.se.exchange.dynamodb.ItemDao;
-import com.nashss.se.exchange.dynamodb.Member;
-import com.nashss.se.exchange.dynamodb.MemberDao;
+import com.nashss.se.exchange.Item;
+import com.nashss.se.exchange.ItemDao;
+import com.nashss.se.exchange.Member;
+import com.nashss.se.exchange.MemberDao;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
