+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/activity/UpdateItemActivity.java
@@ -4,8 +4,8 @@ import com.nashss.se.exchange.Models.ItemModel;
 import com.nashss.se.exchange.activity.requests.UpdateItemRequest;
 import com.nashss.se.exchange.activity.results.UpdateItemResult;
 import com.nashss.se.exchange.converters.ModelConverter;
-import com.nashss.se.exchange.dynamodb.Item;
-import com.nashss.se.exchange.dynamodb.ItemDao;
+import com.nashss.se.exchange.Item;
+import com.nashss.se.exchange.ItemDao;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import utils.XchangeServiceUtils;
