+++ xchangeServiceLambda/src/main/java/com/nashss/se/exchange/dependency/DaoModule.java
@@ -1,5 +1,5 @@
 package com.nashss.se.exchange.dependency;
-import com.nashss.se.exchange.dynamodb.DynamoDbClientProvider;
+import com.nashss.se.exchange.DynamoDbClientProvider;
 
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
